requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
openai>=1.10.0
fpdf2>=2.7.0
python-dotenv>=1.0.0
pydantic>=2.0.0
